# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: 影策 HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /official-assets:
    get:
      tags: [assets]
      summary: 已发布的官方资产目录
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: kind
          schema: { type: string, enum: [image, video, audio] }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: favorite
          schema: { type: integer, enum: [1] }
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
  /official-assets/{id}:
    get:
      tags: [assets]
      summary: 官方资产详情
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "404":
          $ref: "#/components/responses/Error"
  /official-assets/{id}/favorite:
    put:
      tags: [assets]
      summary: 收藏官方资产
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
    delete:
      tags: [assets]
      summary: 取消收藏官方资产
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
  /official-assets/{id}/materialize:
    post:
      tags: [assets]
      summary: 将官方资产加入当前用户素材库
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "404":
          $ref: "#/components/responses/Error"
  /official-assets/{id}/media/{mediaId}/file:
    get:
      tags: [assets]
      summary: 读取有权访问的官方资产媒体文件
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: mediaId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 文件流
        "404":
          $ref: "#/components/responses/Error"
  /admin/official-assets:
    get:
      tags: [admin, assets]
      summary: 管理员官方资产目录
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "403":
          $ref: "#/components/responses/Error"
    post:
      tags: [admin, assets]
      summary: 创建官方资产
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveOfficialAssetRequest"
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
  /admin/official-assets/{id}:
    patch:
      tags: [admin, assets]
      summary: 更新官方资产元数据或未使用资产的媒体
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveOfficialAssetRequest"
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
  /admin/official-assets/{id}/status:
    patch:
      tags: [admin, assets]
      summary: 发布、转为草稿或下架官方资产
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [draft, published, retired]
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
  schemas:
    SaveOfficialAssetRequest:
      type: object
      required: [title, category, status]
      properties:
        title:
          type: string
          maxLength: 120
        category:
          type: string
        status:
          type: string
          enum: [draft, published, retired]
        description:
          type: string
        prompt:
          type: string
        tags:
          type: array
          maxItems: 24
          items: { type: string }
        source:
          type: string
        license:
          type: string
        aiGenerated:
          type: boolean
        sortOrder:
          type: integer
        resourceIds:
          type: array
          minItems: 1
          maxItems: 12
          items: { type: string }
...
    Envelope:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          description: 0 成功；失败时默认等于 HTTP 状态，或 status*100+n 的细分码
        data: {}
        msg:
          type: string
        reason:
          type: string
          description: 仅失败时出现。稳定机器可读原因。
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            code:
              type: integer
              enum: [400, 401, 403, 404, 409, 429, 500, 502, 503, 504, 40301, 40901, 42901]
            data:
              nullable: true
            reason:
              type: string
              enum:
                - invalid_argument
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - failed_precondition
                - quota_exceeded
                - rate_limited
                - unavailable
                - timeout
                - internal
                - bad_gateway
                - model_capability_not_supported
                - model_price_not_configured
                - model_route_unavailable
                - provider_request_failed
                - model_catalog_mismatch
                - invalid_model_selection
  responses:
    Envelope:
      description: 业务成功信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Error:
      description: 业务失败信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"